- Moved types related to svgmapper into shared folder, for future probable usage in client and server
